DRY up guards
