+++ packages/liveblocks-core/src/index.ts
@@ -122,6 +122,7 @@ export type {
 export { makeEventSource } from "./lib/EventSource";
 export * as console from "./lib/fancy-console";
 export { freeze } from "./lib/freeze";
+export { isPlainObject, isStartsWith } from "./lib/guards";
 export type { Json, JsonArray, JsonObject, JsonScalar } from "./lib/Json";
 export { isJsonArray, isJsonObject, isJsonScalar } from "./lib/Json";
 export { nanoid } from "./lib/nanoid";
@@ -142,7 +143,6 @@ export type { Brand, DistributiveOmit } from "./lib/utils";
 export {
   b64decode,
   compactObject,
-  isPlainObject,
   mapValues,
   memoizeOnSuccess,
   raise,
