+++ packages/liveblocks-core/src/lib/utils.ts
@@ -18,18 +18,6 @@ export function raise(msg: string): never {
   throw new Error(msg);
 }
 
-export function isPlainObject(
-  blob: unknown
-): blob is { [key: string]: unknown } {
-  // Implementation borrowed from pojo decoder, see
-  // https://github.com/nvie/decoders/blob/78849f843193647eb6b5307240387bdcff7161fb/src/lib/objects.js#L10-L41
-  return (
-    blob !== null &&
-    typeof blob === "object" &&
-    Object.prototype.toString.call(blob) === "[object Object]"
-  );
-}
-
 /**
  * Drop-in replacement for Object.entries() that retains better types.
  */
