+++ packages/liveblocks-react/src/lib/querying.ts
@@ -1,7 +1,7 @@
 import type { BaseMetadata, ThreadData } from "@liveblocks/client";
+import { isStartsWith } from "@liveblocks/core";
 
 import type { ThreadsQuery } from "../types";
-import { isStartsWith, isString } from "./guards";
 
 /**
  * Creates a predicate function that will filter all ThreadData instances that
@@ -47,7 +47,7 @@ function matchesOperator(
   op: BaseMetadata[string] | { startsWith: string } | undefined
 ) {
   if (isStartsWith(op)) {
-    return isString(value) && value.startsWith(op.startsWith);
+    return typeof value === "string" && value.startsWith(op.startsWith);
   } else {
     return value === op;
   }
